refactor: move ipc message interfaces to `Domain/Common`
