+++ GlazeWM.Application.IpcServer/IpcMessageHandler.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Text.Json;
 using CommandLine;
-using GlazeWM.Application.IpcServer.Messages;
 using GlazeWM.Application.IpcServer.Server;
+using GlazeWM.Domain.Common.IpcMessages;
 using GlazeWM.Domain.Containers;
 using GlazeWM.Domain.Monitors;
 using GlazeWM.Domain.UserConfigs;
