+++ GlazeWM.Application/Program.cs
@@ -5,11 +5,11 @@ using System.Threading;
 using CommandLine;
 using GlazeWM.Application.CLI;
 using GlazeWM.Application.IpcServer;
-using GlazeWM.Application.IpcServer.Messages;
 using GlazeWM.Application.WM;
 using GlazeWM.Bar;
 using GlazeWM.Domain;
 using GlazeWM.Domain.Common;
+using GlazeWM.Domain.Common.IpcMessages;
 using GlazeWM.Domain.Containers;
 using GlazeWM.Infrastructure;
 using GlazeWM.Infrastructure.Bussing;
