+++ GlazeWM.Domain/Common/IpcMessages/GetMonitorsMessage.cs
@@ -1,6 +1,6 @@
 using CommandLine;
 
-namespace GlazeWM.Application.IpcServer.Messages
+namespace GlazeWM.Domain.Common.IpcMessages
 {
   [Verb("monitors", HelpText = "Get all monitors.")]
   public class GetMonitorsMessage
