+++ GlazeWM.Domain/Common/IpcMessages/GetWorkspacesMessage.cs
@@ -1,6 +1,6 @@
 using CommandLine;
 
-namespace GlazeWM.Application.IpcServer.Messages
+namespace GlazeWM.Domain.Common.IpcMessages
 {
   [Verb("workspaces", HelpText = "Get all workspaces.")]
   public class GetWorkspacesMessage
