+++ GlazeWM.Domain/Common/IpcMessages/InvokeCommandMessage.cs
@@ -1,6 +1,6 @@
 using CommandLine;
 
-namespace GlazeWM.Application.IpcServer.Messages
+namespace GlazeWM.Domain.Common.IpcMessages
 {
   [Verb(
     "command",
